requirements.txt
yfinance==0.2.18
pandas==2.0.0
apache-airflow==2.5.3
google-cloud-bigquery==3.10.0